+++ packages/plugin-button/web/src/viewer.js
@@ -1,12 +1,9 @@
 import { typeMapper } from './typeMapper';
-import { BUTTON_TYPE } from './constants';
+import { BUTTON_TYPE, DEFAULT_CONFIG } from './constants';
 export { typeMapper as buttonTypeMapper, BUTTON_TYPE };
-
-const defaultConfig = {};
-
 export const pluginButton = (config = {}) => {
   return {
-    config: { ...defaultConfig, ...config },
+    config: { ...DEFAULT_CONFIG, ...config },
     type: BUTTON_TYPE,
     typeMapper,
     decorator: {},