+++ packages/plugin-html/web/src/viewer.js
@@ -1,10 +1,11 @@
 import { typeMapper } from './typeMapper';
 import { HTML_TYPE } from './types';
+import { DEFAULTS } from './constants';
 export { typeMapper as htmlTypeMapper, HTML_TYPE };
 
 export const pluginHtml = (config = {}) => {
   return {
-    config,
+    config: { ...DEFAULTS.config, ...config },
     type: HTML_TYPE,
     typeMapper,
     decorator: {},