+++ packages/plugin-image/web/src/viewer.js
@@ -1,12 +1,11 @@
 import { typeMapper } from './typeMapper';
 import { IMAGE_TYPE } from './types';
+import { DEFAULTS } from './consts';
 export { typeMapper as imageTypeMapper, IMAGE_TYPE };
-
-const defaultConfig = {};
 
 export const pluginImage = (config = {}) => {
   return {
-    config: { ...defaultConfig, ...config },
+    config: { ...DEFAULTS.config, ...config },
     type: IMAGE_TYPE,
     typeMapper,
     decorator: {},