+++ packages/plugin-video/web/src/video-component.jsx
@@ -64,7 +64,14 @@ class VideoComponent extends React.Component {
   };
 
   renderPlayer = () => {
-    const { theme, componentData, disabled, disableRightClick, settings } = this.props;
+    const {
+      theme,
+      componentData,
+      disabled,
+      disableRightClick,
+      settings,
+      setComponentUrl,
+    } = this.props;
     return (
       <VideoViewer
         ref={this.setPlayer}
@@ -74,6 +81,7 @@ class VideoComponent extends React.Component {
         disabled={disabled}
         disableRightClick={disableRightClick}
         theme={theme}
+        setComponentUrl={setComponentUrl}
       />
     );
   };
@@ -118,6 +126,7 @@ VideoComponent.propTypes = {
   t: PropTypes.func.isRequired,
   disableRightClick: PropTypes.bool,
   disabled: PropTypes.bool,
+  setComponentUrl: PropTypes.func,
 };
 
 export { VideoComponent as Component, DEFAULTS };