+++ packages/wrapper/web/src/PluginsStrategyViewer/index.js
@@ -1,2 +1,0 @@
-import pluginsStrategyProviderViewer from '../PluginsStrategy/PluginsStrategyProviderViewer';
-export default pluginsStrategyProviderViewer;